acpi_call-kmp-default package should be additionally installed on Tumbleweed
Actually, I don't know how to correctly state this in the documentation, but for battery recalibration to work on openSUSE Tumbleweed, the acpi_call-kmp-default package should be installed together with the acpi_call package. The acpi_call-kmp-default package installs the kernel module, while the acpi_call installs the files required to automatically load this kernel module at boot time. This is for Tumbleweed only (acpi_call-kmp-default package isn't available for Leap).